+++ migrations/000004_create_portfolio_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE portfolio (
     id BIGSERIAL PRIMARY KEY,
     user_id BIGINT NOT NULL,
     asset VARCHAR(20) NOT NULL,
-    quantity NUMERIC(20,8) NOT NULL,
+    quantity BIGINT NOT NULL,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     CONSTRAINT fk_portfolio_user
         FOREIGN KEY (user_id)
